chemical-visualizer/backend/requirements.txt
django
djangorestframework
pandas
reportlab
django-cors-headers
gunicorn
whitenoise
dj-database-url
psycopg2-binary
